pyproject.toml
[tool.poetry]
name = "ddb-python-fastapi-demo"
version = "0.1.0"
description = "demo for a python/fastapi/poetry-ddb project"
authors = ["Valentin Gillet <valentin.gillet@inetum.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.115.4"
uvicorn = "^0.32.0"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
